Change data types
This is going to a size_t, when switching from using DllImport to
DynamicInvoke, this stopped working on 32bit systems (though 64 worked
fine), as far as I can tell, this works on both.

+++ libsodium-net/Utilities.cs
@@ -140,7 +140,7 @@ namespace Sodium
     }
 
     //sodium_bin2hex
-    private delegate IntPtr _Bin2Hex(byte[] hex, long hexMaxlen, byte[] bin, long binLen);
+    private delegate IntPtr _Bin2Hex(byte[] hex, int hexMaxlen, byte[] bin, int binLen);
 
     //sodium_hex2bin
     private delegate int _Hex2Bin(IntPtr bin, int binMaxlen, string hex, int hexLen, string ignore, out int binLen, string hexEnd);
